all: rename lime-defauts -> lime-community
In order to convert community profiles into packages a new config file
lime-community was introduced as lime-defaults cannot be used for that.
This is because lime-defaults is already being shipped with the
lime-system package and hence cannot be installed again by any other
package. Hence rename lime-defaults to lime-community in all profiles.

Signed-off-by: Daniel Golle <daniel@makrotopia.org>